+++ Sources/BFKit/Linux/Foundation/DictionaryExtension.swift
@@ -36,11 +36,22 @@ public extension Dictionary {
     ///
     /// - Parameter key: Key to get value of.
     /// - Returns: Value for the key Or nil.
+    @available(*, deprecated: 3.0, message: "Use the `[]` subscript instead.")
     public func safeObject(forKey key: Key) -> Any? {
         if let value = self[key] {
             return value
         } else {
             return nil
         }
+    }
+    
+    /// Append a Value for a given Key in the Dictionary.
+    /// If the Key already exist it will be ovrewritten.
+    ///
+    /// - Parameters:
+    ///   - value: Value to be added.
+    ///   - key: Key to be added.
+    public mutating func append(_ value: Value, forKey key: Key) {
+        self[key] = value
     }
 }