+++ CHANGELOG.md
@@ -36,6 +36,7 @@ All notable changes to this project will be documented in this file.<br>
 - Added iPad Pro 12.9 2nd Generation and iPad Pro 10.5 devices in UIDevice extension
 - Added `fade(label:, duration: = default, toString:)` and `fade(label:, duration: = default, toAttributedString:)` animation functions in WKInterfaceController extension
 - Added `Set` struct extension
+- Added `append(_ value:, key:)` function in Dictionary extension
 
 ### Improved
 - Now `dateString(format: = default)` in Date extension has a new `locale` parameter with a `default` value, so no changes are required